Stats for materials (bronze, steel, etc.) can be defined in datapacks by writing a material.json file (ex: namespace/materials/bronze.json)
Alloygery provides a material registry for other mods to register custom materials. Materials registered this way will be synced between server and client, and clients will respect the material stats defined on the server opposed to their own.

As a cost to moving material stats to datapacks, the config has lost a lot of its customization, but it is worth it to have server defined materials.

Material stats have been rebalanced to assume the presence of vanilla equipment like diamond and netherite, and alloygery no longer overrides any vanilla items. This means that players can craft and use diamond and netherite gear by default, but also means that you can not adjust the stats of vanilla materials like iron, and gold.