Part based tools
Full implementation of part based tools. Vanilla type tools are crafted from a tool head, binding, and a handle. Alloygery tools will have stats and renderings derived from their combination of parts.

The smithing anvil was overhauled to support variable material cost for recipes

This version is not backward compatible. No tool or armor items will survive an upgrade from a precious version.

Known issues:
The crafting recipe in REI for tools shows the error color for the resultant tool, this is a bug with nbt-crafting, and I am still working out a workaround.